[VC] Split Parser into partial files
Move and extend parsing logic into three new partial files and update Parser.cs to use them. Added server/GSCode.Parser/AST/Parser.Declarations.cs, Parser.Expressions.cs and Parser.Statements.cs implementing comprehensive parsing routines: dependency and path parsing, script/namespace/class/function/constructor/destructor declarations, parameter lists, precache and usingAnimTree directives, dev-blocks, class body/member parsing, expression parsing with full operator precedence (assignments, logical/bitwise/arithmetic, shifts, unary, calls/accessors/dereferences, array/indexers, waittill/called-on semantics), statement/brace block handling, error recovery and folding range emission. These changes reorganize the parser into focused partials and implement many AST node constructors and recovery/error reporting behaviors.
